backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cfgv==3.4.0
click==8.1.8
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
flask-mongoengine==1.0.0
Flask-WTF==1.2.2
httplib2==0.20.2
identify==2.6.8
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
mongoengine==0.29.1
more-itertools==8.10.0
netifaces==0.11.0
nodeenv==1.9.1
numpy==2.2.2
oauthlib==3.2.0
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.1.0
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.11.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==5.4.1
ruff==0.9.9
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tzdata==2025.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.29.2
wadllib==1.3.6
Werkzeug==3.1.3
WTForms==3.2.1
zipp==1.0.0
